requirements.txt
Flask
tensorflow
Pillow
numpy
matplotlib
opencv-python
werkzeug
uuid